move BookSide into its own file
